rename organization in authors
