+++ src/config/base.js
@@ -1,4 +1,5 @@
 import build from 'buildType';
+
 export default {
   apiUrl: '',
   homePath: '/',
